+++ src/App.css
@@ -10,7 +10,7 @@
 }
 
 .page {
-    height: 100vh;
+    height: 100%;
 }
 
 .bottom {
